+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+2.2.13
+++++++
+* `vm encryption enable`: now validates --disk encryption keyvault and that key encryption keyvault exists, in line with `vmss encryption enable`. Exposes --force. For more info: see issue #8110
+
 2.2.12
 ++++++
 * 'vm disk/unmanaged-disk' and 'vmss disk': parameter options are more consistent.